requirements.txt
alembic==1.10.3
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-editor==1.0.4
six==1.16.0
SQLAlchemy==2.0.9
Werkzeug==2.2.3
python-jose==3.3.0
python-dotenv==1.0.0
